refactor: rearrange extractors
